+++ packages/browser/src/index.ts
@@ -1,42 +1,4 @@
-export {
-  Breadcrumb,
-  Request,
-  SdkInfo,
-  Event,
-  Exception,
-  Response,
-  Severity,
-  StackFrame,
-  Stacktrace,
-  Status,
-  Thread,
-  User,
-} from '@sentry/types';
-
-export {
-  addGlobalEventProcessor,
-  addBreadcrumb,
-  captureException,
-  captureEvent,
-  captureMessage,
-  configureScope,
-  getHubFromCarrier,
-  getCurrentHub,
-  Hub,
-  Scope,
-  setContext,
-  setExtra,
-  setExtras,
-  setTag,
-  setTags,
-  setUser,
-  withScope,
-} from '@sentry/core';
-
-export { BrowserOptions } from './backend';
-export { BrowserClient, ReportDialogOptions } from './client';
-export { defaultIntegrations, forceLoad, init, lastEventId, onLoad, showReportDialog, flush, close, wrap } from './sdk';
-export { SDK_NAME, SDK_VERSION } from './version';
+export * from './exports';
 
 import { Integrations as CoreIntegrations } from '@sentry/core';
 import { getGlobalObject } from '@sentry/utils';